+++ Editor/Content/GeometryImportSettings.cs
@@ -1,0 +1,102 @@
+﻿using Editor.Common;
+
+namespace Editor.Content
+{
+    public class GeometryImportSettings : ViewModelBase
+    {
+        private bool _calculateNormals;
+        private bool _calculateTangents;
+        private float _smoothingAngle;
+        private bool _reverseHandedness;
+        private bool _importEmbeddedTextures;
+        private bool _importAnimations;
+
+        public bool CalculateNormals
+        {
+            get => _calculateNormals;
+            set
+            {
+                if (_calculateNormals != value)
+                {
+                    _calculateNormals = value;
+                    OnPropertyChanged(nameof(CalculateNormals));
+                }
+            }
+        }
+
+        public bool CalculateTangents
+        {
+            get => _calculateTangents;
+            set
+            {
+                if (_calculateTangents != value)
+                {
+                    _calculateTangents = value;
+                    OnPropertyChanged(nameof(CalculateTangents));
+                }
+            }
+        }
+
+        public float SmoothingAngle
+        {
+            get => _smoothingAngle;
+            set
+            {
+                if (_smoothingAngle != value)
+                {
+                    _smoothingAngle = value;
+                    OnPropertyChanged(nameof(SmoothingAngle));
+                }
+            }
+        }
+
+        public bool ReverseHandedness
+        {
+            get => _reverseHandedness;
+            set
+            {
+                if (_reverseHandedness != value)
+                {
+                    _reverseHandedness = value;
+                    OnPropertyChanged(nameof(ReverseHandedness));
+                }
+            }
+        }
+
+        public bool ImportEmbeddedTextures
+        {
+            get => _importEmbeddedTextures;
+            set
+            {
+                if (_importEmbeddedTextures != value)
+                {
+                    _importEmbeddedTextures = value;
+                    OnPropertyChanged(nameof(ImportEmbeddedTextures));
+                }
+            }
+        }
+
+        public bool ImportAnimations
+        {
+            get => _importAnimations;
+            set
+            {
+                if (_importAnimations != value)
+                {
+                    _importAnimations = value;
+                    OnPropertyChanged(nameof(ImportAnimations));
+                }
+            }
+        }
+
+        public GeometryImportSettings()
+        {
+            CalculateNormals = false;
+            CalculateTangents = false;
+            SmoothingAngle = 178f;
+            ReverseHandedness = false;
+            ImportEmbeddedTextures = true;
+            ImportAnimations = true;
+        }
+    }
+}